clean up auth route
