+++ package.json
@@ -71,6 +71,8 @@
     "react-dom": "0.14.3",
     "react-redux": "^4.0.0",
     "redux": "^3.0.4",
+    "request": "2.67.0",
+    "request-promise": "1.0.2",
     "validator": "4.4.0",
     "winston": "2.1.1"
   }
