+++ server/services/user/userService.js
@@ -5,11 +5,9 @@ import _ from 'lodash';
 import assert from 'assert';
 import bodyParser from 'body-parser';
 import express from 'express';
-import jwt from 'jsonwebtoken';
 import moment from 'moment';
 import mongoose from 'mongoose';
 import q from 'q';
-import url from 'url';
 import validator from 'validator';
 
 // Locals
@@ -17,16 +15,15 @@ import fractionErrors from './../../utils/fractionErrors';
 import middlewareErrors from './../../middleware/errorHandler';
 import middlewareAuth from './../../middleware/tokenAuth';
 import serviceRegistry  from './../serviceRegistry';
+// DB Models
 import User from './userModel';
+
 
 // Use Q promises
 mongoose.Promise = require('q').Promise;
 
 
 // Constants
-
-const FRACTION_TOKEN_SECRET = process.config.fraction.tokenSecret;
-const FRACTION_TOKEN_ISSUER = process.config.fraction.clientId;
 
 // Naming
 const SVC_NAME = 'user';
@@ -35,7 +32,6 @@ const SVC_BASE_URL = serviceRegistry.registry.apis.baseV1 + '/user';
 // Routes
 const ROUTE_CREATE_USER = '/';
 const ROUTE_UPDATE_USER = '/:userId';
-const ROUTE_LOG_IN_USER = '/login';
 
 
 // Service setup
@@ -53,24 +49,6 @@ let router = express.Router();
 router.use(bodyParser.urlencoded({ extended: true }));
 // parse application/json
 router.use(bodyParser.json());
-
-
-// Private Helpers
-
-/**
- * Generate a signed jwt for a user
- *
- * @returns {token} string New signed web token
- */
-let generateToken = function() {
-  let now = moment.utc();
-  let payload = {
-    iss: FRACTION_TOKEN_ISSUER,
-    exp: moment(now).add(1, 'day').utc().valueOf(),
-    iat: now.valueOf()
-  };
-  return jwt.sign(payload, FRACTION_TOKEN_SECRET);
-}
 
 
 // Public API Functions
@@ -135,7 +113,7 @@ function createUser(req, res) {
       // TODO(gavin): ENABLE EMAIL VERIFICATION
       verified: true,
       verifyCode: '123',
-      verifySentAt: new Date().toString()
+      verifySentAt: moment.utc().valueOf()
     },
     local: {
       password: hashedPassword
@@ -144,18 +122,19 @@ function createUser(req, res) {
     fractionEmployee: false
   };
 
-  // TODO(gavin): possibly search for a user first by email as well
-  // currently only enforcing it in the schema
-
-  return User.create(pendingUser)
+  return User.findOne({ 'email.email': email })
+    .then((user) => {
+      if (user) {
+        throw new fractionErrors.Forbidden('user exists');
+      }
+      return User.create(pendingUser);
+    })
     .then((newUser) => {
       return res.json({ user: newUser.toPublicObject() });
     })
     .catch((err) => {
-      // Note: E11000 is mongo's duplicate key error
-      // TODO(gavin): Pull out error handling and messages into universal object
-      if (_.contains(err.message, 'E11000')) {
-        throw new fractionErrors.Forbidden('user exists');
+      if (err instanceof fractionErrors.Forbidden) {
+        throw err;
       }
       throw new Error(err);
     });
@@ -245,8 +224,6 @@ function updateUser(req, res) {
       return existingUser.save();
     })
     .then((updatedUser) => {
-      // console.log('UPDATED: ', updatedUser);
-
       return res.json({ user: updatedUser.toPublicObject() });
     })
     .catch((err) => {
@@ -259,65 +236,10 @@ function updateUser(req, res) {
 };
 
 
-/**
- * Authenticate a user
- *
- * @param {req} obj Express request object
- * @param {res} obj Express response object
- * @returns {promise}
- */
-function logInUser(req, res) {
-
-  let email;
-  let hashedPassword;
-
-  // validate email
-  try {
-    assert(_.isString(req.body.email));
-    email = validator.toString(req.body.email).toLowerCase();
-    assert(validator.isEmail(email));
-  } catch(e) {
-    throw new fractionErrors.Invalid('invalid email');    
-  }
-
-  // password
-  try {
-    assert(_.isString(req.body.password));
-    hashedPassword = validator.toString(req.body.password);
-    assert(hashedPassword.length);
-  } catch(e) {
-    throw new fractionErrors.Invalid('invalid password');
-  }
-
-  return User.findOne({ 'email.email': email, 'local.password': hashedPassword })
-    .then((user) => {
-      if (!user) {
-        throw new fractionErrors.NotFound('user not found');
-      }
-
-      // Generate a new token for the user
-      let token;
-      try {
-        token = generateToken(req);
-      } catch (err) {
-        throw new Error('error generating user token');
-      }
-      return res.json({ token: token, user: user.toPublicObject() });
-    })
-    .catch((err) => {
-      if (err instanceof fractionErrors.NotFound) {
-        throw err;
-      }
-      throw new Error(err.message);
-    });
-};
-
-
 // Routes
 
 router.post(ROUTE_CREATE_USER, middlewareErrors.wrap(createUser));
 router.put(ROUTE_UPDATE_USER, middlewareAuth.requireAuth, middlewareErrors.wrap(updateUser));
-router.post(ROUTE_LOG_IN_USER, middlewareErrors.wrap(logInUser));
 
 
 // Exports
@@ -328,29 +250,9 @@ module.exports = {
   router: router,
   endpoints: [
     { protocol: 'HTTP', method: 'POST', name: 'CREATE_USER', url: ROUTE_CREATE_USER },
-    { protocol: 'HTTP', method: 'PUT', name: 'UPDATE_USER', url: ROUTE_UPDATE_USER },
-    { protocol: 'HTTP', method: 'POST', name: 'LOG_IN_USER', url: ROUTE_LOG_IN_USER }
+    { protocol: 'HTTP', method: 'PUT', name: 'UPDATE_USER', url: ROUTE_UPDATE_USER }
   ]
 };
 
 // Register with the app service registry
 serviceRegistry.registry.register(module.exports);
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
